pr comments and fix imports
