+++ packages/taquito-contracts-library/src/taquito-contracts-library.ts
@@ -6,7 +6,8 @@
 import { EntrypointsResponse, ScriptedContracts } from '@taquito/rpc';
 import { Extension, Context } from '@taquito/taquito';
 import { validateAddress, ValidationResult } from '@taquito/utils';
-import { InvalidScriptFormatError, InvalidAddressError } from './errors';
+import { InvalidContractAddressError } from '@taquito/core';
+import { InvalidScriptFormatError } from './errors';
 import { ReadWrapperContractsLibrary } from './read-provider-wrapper';
 
 interface ContractsData {
@@ -70,7 +71,7 @@ export class ContractsLibrary implements Extension {
 
   private validateContractAddress(address: string) {
     if (validateAddress(address) !== ValidationResult.VALID) {
-      throw new InvalidAddressError(address);
+      throw new InvalidContractAddressError(address);
     }
   }
