+++ packages/taquito-core/src/error/common.ts
@@ -32,3 +32,14 @@ export class InvalidHexStringError extends Error {
     super(message);
   }
 }
+
+/**
+ *  @category Error
+ *  @description Error that indicates an invalid contract address being passed or used
+ */
+export class InvalidContractAddressError extends Error {
+  public name = 'InvalidContractAddressError';
+  constructor(public contractAddress: string) {
+    super(`The contract address '${contractAddress}' is invalid`);
+  }
+}
