+++ packages/taquito-ledger-signer/src/taquito-ledger-signer.ts
@@ -6,6 +6,7 @@
 import { Signer } from '@taquito/taquito';
 import Transport from '@ledgerhq/hw-transport';
 import { b58cencode, prefix, Prefix, ProhibitedActionError } from '@taquito/utils';
+import { InvalidDerivationPathError } from '@taquito/core';
 import {
   appendWatermark,
   transformPathToBuffer,
@@ -19,7 +20,7 @@ import {
   PublicKeyHashRetrievalError,
   PublicKeyRetrievalError,
   InvalidLedgerResponseError,
-  InvalidDerivationPathError,
+  InvalidDerivationTypeError,
 } from './error';
 
 export type LedgerTransport = Pick<Transport, 'send' | 'decorateAppAPIMethods' | 'setScrambleKey'>;
@@ -31,18 +32,6 @@ export enum DerivationType {
   BIP32_ED25519 = 0x03, // tz1 BIP32
 }
 
-/**
- *  @category Error
- *  @description Error that indicates an invalid derivation type being passed or used
- */
-export class InvalidDerivationTypeError extends Error {
-  public name = 'InvalidDerivationTypeError';
-  constructor(public derivationType: string) {
-    super(
-      `The derivation type ${derivationType} is invalid. The derivation type must be DerivationType.ED25519, DerivationType.SECP256K1, DerivationType.P256 or DerivationType.BIP32_ED25519`
-    );
-  }
-}
 
 export const HDPathTemplate = (account: number) => {
   return `44'/1729'/${account}'/0'`;
