+++ packages/taquito-local-forging/src/michelson/codec.ts
@@ -4,7 +4,8 @@ import { Uint8ArrayConsumer } from '../uint8array-consumer';
 import { Encoder } from '../taquito-local-forging';
 import { opMappingReverse, opMapping } from '../constants';
 import { pad } from '../utils';
-import { UnexpectedMichelsonValueError, InvalidHexStringError } from '../error';
+import { UnexpectedMichelsonValueError } from '../error';
+import { InvalidHexStringError } from '@taquito/core'
 
 export type PrimValue = { prim: string; args?: MichelsonValue[]; annots?: string[]; };
 export type BytesValue = { bytes: string; };
