+++ packages/taquito-local-forging/src/taquito-local-forging.ts
@@ -3,14 +3,13 @@
  * @module @taquito/local-forging
  */
 
-import { InvalidBlockHashError } from './error';
 import { ForgeParams, Forger } from './interface';
 import { CODEC } from './constants';
 import { decoders } from './decoder';
 import { encoders } from './encoder';
 import { Uint8ArrayConsumer } from './uint8array-consumer';
 import { validateBlock, ValidationResult, InvalidOperationKindError } from '@taquito/utils';
-import { InvalidOperationSchemaError } from './error';
+import { InvalidOperationSchemaError, InvalidBlockHashError } from './error';
 import { validateMissingProperty, validateOperationKind } from './validator';
 import { ProtocolsHash, ProtoInferiorTo } from './protocols';
 import { decodersProto14 } from './proto14-kathmandu/decoder-proto14';
