+++ packages/taquito-local-forging/src/uint8array-consumer.ts
@@ -1,4 +1,4 @@
-import { InvalidHexStringError } from './error';
+import { InvalidHexStringError } from '@taquito/core';
 
 export class Uint8ArrayConsumer {
   static fromHexString(hex: string) {
@@ -13,7 +13,7 @@ export class Uint8ArrayConsumer {
     }
   }
 
-  constructor(private readonly arr: Uint8Array, private offset: number = 0) { }
+  constructor(private readonly arr: Uint8Array, private offset: number = 0) {}
 
   public consume(count: number): Uint8Array {
     const subArr = this.arr.subarray(this.offset, this.offset + count);
