+++ packages/taquito-rpc/src/rpc-client-modules/rpc-cache.ts
@@ -47,14 +47,12 @@ import {
   VotesListingsResponse,
   VotingPeriodBlockResult,
 } from '../types';
-
 import {
-  InvalidContractAddressError,
   validateContractAddress,
   validateAddress,
   ValidationResult,
 } from '@taquito/utils';
-import { InvalidAddressError } from '../errors';
+import { InvalidContractAddressError, InvalidAddressError } from '@taquito/core';
 
 interface CachedDataInterface {
   [key: string]: {
