+++ packages/taquito-sapling/src/taquito-sapling.ts
@@ -14,7 +14,8 @@ import {
   validateKeyHash,
   ValidationResult,
 } from '@taquito/utils';
-import { InsufficientBalance, InvalidMemo, InvalidAddressError } from './error';
+import { InsufficientBalance, InvalidMemo } from './error';
+import { InvalidAddressError } from '@taquito/core';
 import { convertValueToBigNumber } from './sapling-tx-viewer/helpers';
 import { InMemorySpendingKey } from './sapling-keys/in-memory-spending-key';
 import { SaplingForger } from './sapling-forger/sapling-forger';
