+++ packages/taquito-signer/src/derivation-tools/index.ts
@@ -1,4 +1,4 @@
-import { InvalidDerivationPathError } from '../errors';
+import { InvalidDerivationPathError } from '@taquito/core';
 
 export * as ECDSA from "./ecdsa";
 export * as Ed25519 from "./ed25519";
