+++ packages/taquito/src/contract/rpc-contract-provider.ts
@@ -53,10 +53,9 @@ import { smartContractAbstractionSemantic } from './semantic';
 import {
   validateAddress,
   validateContractAddress,
-  InvalidContractAddressError,
   ValidationResult,
 } from '@taquito/utils';
-import { InvalidAddressError } from '@taquito/core';
+import { InvalidAddressError, InvalidContractAddressError } from '@taquito/core';
 import { EstimationProvider } from '../estimate/estimate-provider-interface';
 import { TxRollupOriginationOperation } from '../operations/tx-rollup-origination-operation';
 import { TxRollupBatchOperation } from '../operations/tx-rollup-batch-operation';
