+++ packages/taquito/src/wallet/wallet.ts
@@ -19,11 +19,10 @@ import {
 import {
   validateAddress,
   validateContractAddress,
-  InvalidContractAddressError,
   ValidationResult,
   InvalidOperationKindError,
 } from '@taquito/utils';
-import { InvalidAddressError } from '@taquito/core';
+import { InvalidAddressError, InvalidContractAddressError } from '@taquito/core';
 
 export interface PKHOption {
   forceRefetch?: boolean;
@@ -37,7 +36,7 @@ export type WalletParamsWithKind =
 export class WalletOperationBatch {
   private operations: WalletParamsWithKind[] = [];
 
-  constructor(private walletProvider: WalletProvider, private context: Context) { }
+  constructor(private walletProvider: WalletProvider, private context: Context) {}
 
   /**
    *
@@ -156,7 +155,7 @@ export class WalletOperationBatch {
 }
 
 export class Wallet {
-  constructor(private context: Context) { }
+  constructor(private context: Context) {}
 
   private get walletProvider() {
     return this.context.walletProvider;
