+++ packages/taquito/test/estimate/rpc-estimate-provider.spec.ts
@@ -26,7 +26,7 @@ import {
 } from '../contract/helper';
 import { OpKind } from '@taquito/rpc';
 import { TransferTicketParams } from '../../src/operations/types';
-import { InvalidContractAddressError } from '@taquito/utils';
+import { InvalidContractAddressError } from '@taquito/core';
 import { InvalidAddressError } from '@taquito/core';
